@tailhook 2020-01-19
While I'm yet not sure where the attitude of removing unsafe code everywhere will lead us in general. In this case yes, I think the overhead does not worth optimizing. Having less to worry about for users matters more.

Merged. Thanks!