@fbernier 2022-06-29
FYI, we've had to fork this code in a project where this is called in a hot loop and the 4% speed difference for use is very worth it. I know this type of change is all about tradeoffs but I personally believe this is easy enough to audit and worth optimizing for.